프로필바 FeedPopup 연결, 파일 이름 통합
